pyproject.toml
[project]
name = "bidder"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiogram",
    "python-dotenv",
    "openai",
    "pillow",
    "pydantic",
    "loguru>=0.7.3",
    "apscheduler",
]
